requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
idna==2.7
oauthlib==2.1.0
py-o365==0.1.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
stringcase==1.2.0
tzlocal==1.5.1
urllib3==1.23
